Move tests over from sokol-samples (#666)
This is mainly so that CI works as expected for Github PRs.

Also fixes a couple of uncritical warnings in sokol_app.h, sokol_audio.h and sokol_imgui.h

+++ util/sokol_imgui.h
@@ -2072,8 +2072,8 @@ SOKOL_API_IMPL void simgui_render(void) {
             }
         }
         #if defined(__cplusplus)
-            const size_t vtx_size = cl->VtxBuffer.size() * sizeof(ImDrawVert);
-            const size_t idx_size = cl->IdxBuffer.size() * sizeof(ImDrawIdx);
+            const size_t vtx_size = (size_t)cl->VtxBuffer.size() * sizeof(ImDrawVert);
+            const size_t idx_size = (size_t)cl->IdxBuffer.size() * sizeof(ImDrawIdx);
         #else
             const size_t vtx_size = (size_t)cl->VtxBuffer.Size * sizeof(ImDrawVert);
             const size_t idx_size = (size_t)cl->IdxBuffer.Size * sizeof(ImDrawIdx);
